apps/crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "restate-sdk>=0.12.0",
    "crawl4ai>=0.7.7",
    "pydantic-ai==1.22.0",
    "pydantic-ai-slim[ollama]==1.22.0",
    "hypercorn>=0.16.0",
    "playwright>=1.40.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a27",
    "ruff>=0.1.0",
]


[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []
